// src/app/interfaces/spouse.ts
export interface Spouse {
    idSpouse: any;
    paternalLastName: any;
    maternalLastName: any;
    name: any;
    rfc: any;
    curp: any;
    oldPaternalLastName: any;
    oldMaternalLastName: any;
    oldName: any;
    nationality: any;
    speticNationality: any;
    gender: any;
    countryOfResidence: any;
    dateOfBirth: any;
    creditApplication: any;
}